+++ examples/showoci/CHANGELOG.rst
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 =====================
+22.06.21 - 22.06.21
+=====================
+* Added dbservers id to exadata csv
+
+=====================
+22.05.31 - 22.05.31
+=====================
+* Added Compartment Path to every api and csv
+* Added Visual Builder
+* Added csvcol to extract defined tag to csv columns
+
+=====================
 22.05.24 - 22.05.24
 =====================
 * Added Shape Capacity to Reservation CSV